Add limited support for proper Lazyfication.
We can properly lazify arguments now, by generating a function pointer
to a slice function which computes a given value. This function pointer
is then used to compute the value that would otherwise be eagerly
evaluated. Note that we currently limit the implementation to *only
optimizing slices which do NOT receive input arguments*.